requirements-cpu.txt
# CPU-only requirements for faster installation
torch>=2.0.0+cpu --index-url https://download.pytorch.org/whl/cpu
torch-geometric>=2.3.0

numpy>=1.24
pandas>=2.0
polars>=1.0
sqlalchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9
pyyaml>=6.0
aiohttp>=3.9
aiohttp-sse-client>=0.2.1
pytest-asyncio>=0.23
stellar-sdk>=9.0.0
tenacity>=8.4.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
